requirements.txt
# 에브리타임 크롤링 프로젝트 의존성
# Python 3.9+ 기준

# 웹 자동화
selenium==4.15.2

# HTML 파싱
beautifulsoup4==4.12.2

# HTTP 요청 및 세션 관리
requests==2.31.0

# 환경변수 관리 (ID/PW 숨기기용)
python-dotenv==1.0.0

# 추가 유틸리티 패키지들
# 웹드라이버 자동 관리
webdriver-manager==4.0.1

# 데이터 처리
pandas==2.1.3

# 로깅
loguru==0.7.2

# Flask API 서버
flask==3.0.0
flask-cors==4.0.0
pinecone-client==4.1.0
sentence-transformers==3.1.1
